bau: publish to bintray/alphagov
Once this is merged we will need to update the jenkins job for this repo
so it has BINTRAY_USER and BINTRAY_API_KEY set (securely).

This will mean that we can make the next level of libs build on travis
with the perl hack (or similar) in .travis.yml.

Pushing to the `maven-test` instead of the `maven` repo to dissuade
third parties from using these builds, as I don't think we'd want to
support them.